unified export commands into one including one file
