+++ entityDumper/types.go
@@ -4,18 +4,18 @@ import (
 	"github.com/uyuni-project/inter-server-sync/utils"
 )
 
-type ChannelDumperOptions struct {
+type DumperOptions struct {
 	ServerConfig              string
 	ChannelLabels             []string
+	ConfigLabels              []string
 	ChannelWithChildrenLabels []string
 	OutputFolder              string
 	outputFolderAbsPath       string
 	MetadataOnly              bool
 	StartingDate              string
-	ConfigLabels []string
 }
 
-func (opt *ChannelDumperOptions) GetOutputFolderAbsPath() string {
+func (opt *DumperOptions) GetOutputFolderAbsPath() string {
 	if "" == opt.outputFolderAbsPath {
 		opt.outputFolderAbsPath = utils.GetAbsPath(opt.OutputFolder)
 	}
